+++ src/testing/index.ts
@@ -1,16 +1,3 @@
 import '../utilities/globals';
-export { MockedProvider, MockedProviderProps } from './mocking/MockedProvider';
-export {
-  MockLink,
-  mockSingleLink,
-  MockedResponse,
-  ResultFunction
-} from './mocking/mockLink';
-export {
-  MockSubscriptionLink,
-  mockObservableLink
-} from './mocking/mockSubscriptionLink';
-export { createMockClient } from './mocking/mockClient';
-export { default as subscribeAndCount } from './subscribeAndCount';
-export { itAsync } from './itAsync';
-export * from './withConsoleSpy';
+export { MockedProvider, MockedProviderProps } from './react/MockedProvider';
+export * from './core';